requirements.txt
fastapi==0.115.11
pydantic==2.10.6
pydantic_core==2.27.2
starlette==0.46.1
uvicorn==0.34.0
h11==0.14.0
motor==3.7.0
joblib==1.4.2
python-dotenv==1.0.1
cryptography==44.0.2
numpy==2.2.4
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.6.0
typing_extensions==4.12.2
sniffio==1.3.1
multidict==6.2.0
idna==3.6
urllib3==2.2.1
gunicorn==23.0.0
datetime==5.5
gspread==6.2.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
cryptography==44.0.2
transformers==4.49.0
torch==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
accelerator==2024.9.13